back/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
graphene-django==3.2.3
django-cors-headers==4.3.1
django-environ==0.11.2
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2
factory-boy==3.3.0